requirements.txt
openai
langchain
langchain_openai
langchain_community
langchain_experimental
python-dotenv